requirements.txt
pandas
numpy
streamlit
plotly
google-cloud-bigquery
google-auth